pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "newsprism"
version = "0.1.0"
description = "Multilingual event-centric news monitor that publishes a Chinese cross-source digest"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "NewsPrism contributors" },
]
keywords = [
    "news",
    "aggregation",
    "multilingual",
    "clustering",
    "summarization",
    "media-analysis",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content :: News/Diary",
    "Topic :: Scientific/Engineering :: Information Analysis",
]
dependencies = [
    # Collection
    "feedparser>=6.0",
    "httpx>=0.27",
    "trafilatura>=1.9",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    # AI / ML
    "litellm>=1.30",
    "sentence-transformers>=3.0",
    # Telegram
    "python-telegram-bot>=21.0",
    # Templating
    "jinja2>=3.1",
    # Scheduling
    "apscheduler>=3.10",
    # Config / utils
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "tenacity>=8.2",
    "python-dateutil>=2.9",
    "rapidfuzz>=3.6",
    "numpy>=1.26",
    # DB
    # sqlite3 is stdlib
]

[project.optional-dependencies]
dev = [
    "build>=1.2.2",
    "pytest>=9.0",
]

[project.scripts]
newsprism = "newsprism.__main__:main"

[project.urls]
Homepage = "https://github.com/moguiyu/NewsPrism"
Repository = "https://github.com/moguiyu/NewsPrism"
Issues = "https://github.com/moguiyu/NewsPrism/issues"

[tool.hatch.build.targets.wheel]
packages = ["newsprism"]

[tool.pytest.ini_options]
testpaths = ["tests"]
